pyproject.toml
[project]
name = "ai-mail-relay"
version = "0.1.0"
description = "Fetch arXiv AI emails, summarize with an LLM, and forward a digest."
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
dev = []

[project.scripts]
ai-mail-relay = "ai_mail_relay.main:main"
